+++ mvc/src/main/java/com/interface21/webmvc/servlet/controller/DispatcherServlet.java
@@ -1,7 +1,10 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.controller;
 
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.View;
+import com.interface21.webmvc.servlet.controller.handler.AnnotationHandlerMapping;
+import com.interface21.webmvc.servlet.controller.handler.HandlerAdapter;
+import com.interface21.webmvc.servlet.controller.handler.HandlerExecution;
+import com.interface21.webmvc.servlet.view.ModelAndView;
+import com.interface21.webmvc.servlet.view.View;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
