+++ mvc/src/main/java/com/interface21/webmvc/servlet/controller/handler/HandlerAdapter.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.controller.handler;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
